refactor: ui/preview를 도메인별 7개 파일로 재구조화
거대 PreviewData(600줄+) 단일 object와 timetable 색상 분기 +
search 태그가 섞여있던 PreviewSampleData를 도메인 단위 object로 분리:

- LecturePreviewData (lecture/syllabus/custom + 시간표 색상 분기 + reminder/review)
- ThemePreviewData (custom/market/editing theme + sampleCustomTheme)
- TableSummaryPreviewData (drawer table summaries + courseBooks)
- FriendPreviewData (friend/requested/friendTable)
- NotificationPreviewData
- SearchTagPreviewData
- DiaryPreviewData (그대로)

각 파일이 한 도메인만 다루고 모두 object 형태로 통일.
