+++ app/src/main/java/com/wafflestudio/snutt2/feature/friend/FriendsDrawer.kt
@@ -36,9 +36,10 @@ import com.wafflestudio.snutt2.domain.model.TableTrimParam
 import com.wafflestudio.snutt2.ui.components.compose.RedDot
 import com.wafflestudio.snutt2.ui.components.compose.SnuttIcon
 import com.wafflestudio.snutt2.ui.components.compose.clicks
-import com.wafflestudio.snutt2.ui.preview.PreviewData
+import com.wafflestudio.snutt2.ui.preview.FriendPreviewData
 import com.wafflestudio.snutt2.ui.preview.SnuttPreview
 import com.wafflestudio.snutt2.ui.preview.SnuttPreviewSurface
+import com.wafflestudio.snutt2.ui.preview.TableSummaryPreviewData
 import com.wafflestudio.snutt2.ui.theme.SNUTTColors
 import com.wafflestudio.snutt2.ui.theme.SNUTTTypography
 
@@ -396,12 +397,12 @@ private fun FriendRequestItem(
 }
 
 private val sampleFriendsUiStateLoaded = FriendsUiState.Loaded(
-    activeFriends = PreviewData.sampleFriends,
-    requestedFriends = PreviewData.sampleRequestedFriends,
-    selectedFriend = PreviewData.sampleFriends.firstOrNull(),
-    selectedFriendCourseBooks = PreviewData.sampleCourseBooks,
-    selectedCourseBook = PreviewData.sampleCourseBooks.firstOrNull(),
-    selectedFriendTable = PreviewData.sampleFriendTable,
+    activeFriends = FriendPreviewData.sampleFriends,
+    requestedFriends = FriendPreviewData.sampleRequestedFriends,
+    selectedFriend = FriendPreviewData.sampleFriends.firstOrNull(),
+    selectedFriendCourseBooks = TableSummaryPreviewData.sampleCourseBooks,
+    selectedCourseBook = TableSummaryPreviewData.sampleCourseBooks.firstOrNull(),
+    selectedFriendTable = FriendPreviewData.sampleFriendTable,
     selectedFriendTableTheme = BuiltInTheme.SNUTT,
     selectedFriendTableTrimParam = TableTrimParam.Default,
     drawerTab = FriendDrawerTab.ACTIVE,
