+++ app/src/main/java/com/wafflestudio/snutt2/feature/home/drawer/bottomsheet/SelectThemeSheetContent.kt
@@ -34,9 +34,9 @@ import com.wafflestudio.snutt2.feature.themeconfig.AddThemeItem
 import com.wafflestudio.snutt2.ui.components.compose.ThemeIcon
 import com.wafflestudio.snutt2.ui.components.compose.clicks
 import com.wafflestudio.snutt2.ui.components.compose.displayName
-import com.wafflestudio.snutt2.ui.preview.PreviewData
 import com.wafflestudio.snutt2.ui.preview.SnuttPreview
 import com.wafflestudio.snutt2.ui.preview.SnuttPreviewSurface
+import com.wafflestudio.snutt2.ui.preview.ThemePreviewData
 import com.wafflestudio.snutt2.ui.theme.SNUTTColors
 import com.wafflestudio.snutt2.ui.theme.SNUTTTypography
 import com.wafflestudio.snutt2.ui.theme.isDarkMode
@@ -190,14 +190,14 @@ private fun SelectThemeSheetContent_CustomAndBuiltIn() {
         SelectThemeSheetContent(
             sheetType = HomeDrawerBottomSheetType.SelectTheme(
                 customThemes = listOf(
-                    PreviewData.previewCustomTheme1,
-                    PreviewData.previewCustomTheme2,
+                    ThemePreviewData.previewCustomTheme1,
+                    ThemePreviewData.previewCustomTheme2,
                 ),
                 builtInThemes = listOf(
                     BuiltInTheme.SNUTT,
                     BuiltInTheme.MODERN,
                 ),
-                selectedPreviewTheme = PreviewData.previewCustomTheme1,
+                selectedPreviewTheme = ThemePreviewData.previewCustomTheme1,
             ),
             onClickPreviewTheme = {},
             onApply = {},
