+++ app/src/main/java/com/wafflestudio/snutt2/feature/lecturedetail/LectureActionButtons.kt
@@ -20,7 +20,7 @@ import com.wafflestudio.snutt2.domain.model.Lecture
 import com.wafflestudio.snutt2.domain.model.LectureSyllabusInfo
 import com.wafflestudio.snutt2.domain.model.SyllabusLecture
 import com.wafflestudio.snutt2.ui.components.compose.clicks
-import com.wafflestudio.snutt2.ui.preview.PreviewData
+import com.wafflestudio.snutt2.ui.preview.LecturePreviewData
 import com.wafflestudio.snutt2.ui.preview.SnuttPreview
 import com.wafflestudio.snutt2.ui.preview.SnuttPreviewSurface
 import com.wafflestudio.snutt2.ui.theme.SNUTTColors
@@ -113,7 +113,7 @@ private fun LectureDetailActionButton(
 private fun LectureActionButtons_SyllabusLecture_ViewMode() {
     SnuttPreviewSurface {
         LectureActionButtons(
-            lecture = PreviewData.syllabusLecture,
+            lecture = LecturePreviewData.syllabusLecture,
             editMode = false,
             onSyllabus = {},
             onReview = {},
@@ -128,7 +128,7 @@ private fun LectureActionButtons_SyllabusLecture_ViewMode() {
 private fun LectureActionButtons_SyllabusLecture_EditMode() {
     SnuttPreviewSurface {
         LectureActionButtons(
-            lecture = PreviewData.syllabusLecture,
+            lecture = LecturePreviewData.syllabusLecture,
             editMode = true,
             onSyllabus = {},
             onReview = {},
