+++ app/src/main/java/com/wafflestudio/snutt2/feature/lecturedetail/LectureReviewRatingField.kt
@@ -24,7 +24,7 @@ import androidx.compose.ui.unit.sp
 import com.wafflestudio.snutt2.R
 import com.wafflestudio.snutt2.domain.model.LectureReviewInfo
 import com.wafflestudio.snutt2.ui.components.compose.SnuttIcon
-import com.wafflestudio.snutt2.ui.preview.PreviewData
+import com.wafflestudio.snutt2.ui.preview.LecturePreviewData
 import com.wafflestudio.snutt2.ui.preview.SnuttPreview
 import com.wafflestudio.snutt2.ui.preview.SnuttPreviewSurface
 import com.wafflestudio.snutt2.ui.theme.SNUTTColors
@@ -80,7 +80,7 @@ internal fun LectureReviewRatingField(
 private fun LectureReviewRatingField_WithReviewInfo() {
     SnuttPreviewSurface {
         LectureReviewRatingField(
-            reviewInfo = PreviewData.sampleReviewInfo,
+            reviewInfo = LecturePreviewData.sampleReviewInfo,
         )
     }
 }
