+++ app/src/main/java/com/wafflestudio/snutt2/feature/settings/LectureReminderScreen.kt
@@ -46,7 +46,7 @@ import com.wafflestudio.snutt2.ui.components.compose.snackbar.CustomSnackBarDura
 import com.wafflestudio.snutt2.ui.components.compose.snackbar.CustomSnackBarHost
 import com.wafflestudio.snutt2.ui.components.compose.snackbar.CustomSnackBarHostState
 import com.wafflestudio.snutt2.ui.components.compose.snackbar.dismiss
-import com.wafflestudio.snutt2.ui.preview.PreviewData
+import com.wafflestudio.snutt2.ui.preview.LecturePreviewData
 import com.wafflestudio.snutt2.ui.preview.SnuttPreview
 import com.wafflestudio.snutt2.ui.preview.SnuttPreviewSurface
 import com.wafflestudio.snutt2.ui.theme.SNUTTColors
@@ -324,7 +324,7 @@ fun LectureReminderEmpty() {
 private fun LectureReminderScreen_Success() {
     SnuttPreviewSurface {
         LectureReminderScreen(
-            uiState = LectureReminderUiState.Success(PreviewData.sampleLectureReminderOptions),
+            uiState = LectureReminderUiState.Success(LecturePreviewData.sampleLectureReminderOptions),
             onClickBack = {},
             onChangeReminderOption = { _, _ -> },
         )
