+++ app/src/screenshotTest/java/com/wafflestudio/snutt2/feature/lecturedetail/LectureActionButtonsScreenshotTest.kt
@@ -3,14 +3,14 @@ package com.wafflestudio.snutt2.feature.lecturedetail
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
 import com.android.tools.screenshot.PreviewTest
-import com.wafflestudio.snutt2.ui.preview.PreviewData
+import com.wafflestudio.snutt2.ui.preview.LecturePreviewData
 
 @PreviewTest
 @Preview(showBackground = true, widthDp = 360, locale = "ko")
 @Composable
 fun LectureActionButtons_SyllabusLecture_ViewMode() {
     LectureActionButtons(
-        lecture = PreviewData.syllabusLecture,
+        lecture = LecturePreviewData.syllabusLecture,
         editMode = false,
         onSyllabus = {},
         onReview = {},
@@ -24,7 +24,7 @@ fun LectureActionButtons_SyllabusLecture_ViewMode() {
 @Composable
 fun LectureActionButtons_SyllabusLecture_EditMode() {
     LectureActionButtons(
-        lecture = PreviewData.syllabusLecture,
+        lecture = LecturePreviewData.syllabusLecture,
         editMode = true,
         onSyllabus = {},
         onReview = {},
@@ -38,7 +38,7 @@ fun LectureActionButtons_SyllabusLecture_EditMode() {
 @Composable
 fun LectureActionButtons_CustomLecture_ViewMode() {
     LectureActionButtons(
-        lecture = PreviewData.customLecture,
+        lecture = LecturePreviewData.customLecture,
         editMode = false,
         onSyllabus = {},
         onReview = {},
@@ -52,7 +52,7 @@ fun LectureActionButtons_CustomLecture_ViewMode() {
 @Composable
 fun LectureActionButtons_CustomLecture_EditMode() {
     LectureActionButtons(
-        lecture = PreviewData.customLecture,
+        lecture = LecturePreviewData.customLecture,
         editMode = true,
         onSyllabus = {},
         onReview = {},
@@ -66,7 +66,7 @@ fun LectureActionButtons_CustomLecture_EditMode() {
 @Composable
 fun LectureActionButtons_SearchedLecture_ViewMode() {
     LectureActionButtons(
-        lecture = PreviewData.searchedLecture,
+        lecture = LecturePreviewData.searchedLecture,
         editMode = false,
         onSyllabus = {},
         onReview = {},
@@ -80,7 +80,7 @@ fun LectureActionButtons_SearchedLecture_ViewMode() {
 @Composable
 fun LectureActionButtons_HideDeleteButton_SyllabusLecture_ViewMode() {
     LectureActionButtons(
-        lecture = PreviewData.syllabusLecture,
+        lecture = LecturePreviewData.syllabusLecture,
         editMode = false,
         hideDeleteButton = true,
         onSyllabus = {},
