+++ app/src/screenshotTest/java/com/wafflestudio/snutt2/feature/lecturedetail/LectureDetailInfoFieldsScreenshotTest.kt
@@ -3,14 +3,14 @@ package com.wafflestudio.snutt2.feature.lecturedetail
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
 import com.android.tools.screenshot.PreviewTest
-import com.wafflestudio.snutt2.ui.preview.PreviewData
+import com.wafflestudio.snutt2.ui.preview.LecturePreviewData
 
 @PreviewTest
 @Preview(showBackground = true, widthDp = 360, locale = "ko")
 @Composable
 fun LectureDetailInfoFields_SyllabusLecture_ViewMode_WithCategoryPre2025() {
     LectureDetailInfoFields(
-        lecture = PreviewData.syllabusLecture,
+        lecture = LecturePreviewData.syllabusLecture,
         editMode = false,
         showCategoryPre2025 = true,
         onDepartmentChange = {},
@@ -28,7 +28,7 @@ fun LectureDetailInfoFields_SyllabusLecture_ViewMode_WithCategoryPre2025() {
 @Composable
 fun LectureDetailInfoFields_SyllabusLecture_ViewMode_WithoutCategoryPre2025() {
     LectureDetailInfoFields(
-        lecture = PreviewData.syllabusLecture,
+        lecture = LecturePreviewData.syllabusLecture,
         editMode = false,
         showCategoryPre2025 = false,
         onDepartmentChange = {},
@@ -46,7 +46,7 @@ fun LectureDetailInfoFields_SyllabusLecture_ViewMode_WithoutCategoryPre2025() {
 @Composable
 fun LectureDetailInfoFields_SyllabusLecture_EditMode() {
     LectureDetailInfoFields(
-        lecture = PreviewData.syllabusLecture,
+        lecture = LecturePreviewData.syllabusLecture,
         editMode = true,
         showCategoryPre2025 = true,
         onDepartmentChange = {},
@@ -64,7 +64,7 @@ fun LectureDetailInfoFields_SyllabusLecture_EditMode() {
 @Composable
 fun LectureDetailInfoFields_CustomLecture_ViewMode() {
     LectureDetailInfoFields(
-        lecture = PreviewData.customLecture,
+        lecture = LecturePreviewData.customLecture,
         editMode = false,
         showCategoryPre2025 = false,
         onDepartmentChange = {},
@@ -82,7 +82,7 @@ fun LectureDetailInfoFields_CustomLecture_ViewMode() {
 @Composable
 fun LectureDetailInfoFields_CustomLecture_EditMode() {
     LectureDetailInfoFields(
-        lecture = PreviewData.customLecture,
+        lecture = LecturePreviewData.customLecture,
         editMode = true,
         showCategoryPre2025 = false,
         onDepartmentChange = {},
