+++ app/src/screenshotTest/java/com/wafflestudio/snutt2/feature/lecturedetail/LectureHeaderFieldsScreenshotTest.kt
@@ -4,14 +4,14 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
 import com.android.tools.screenshot.PreviewTest
 import com.wafflestudio.snutt2.domain.model.BuiltInTheme
-import com.wafflestudio.snutt2.ui.preview.PreviewData
+import com.wafflestudio.snutt2.ui.preview.LecturePreviewData
 
 @PreviewTest
 @Preview(showBackground = true, widthDp = 360, locale = "ko")
 @Composable
 fun LectureHeaderFields_LectureUIInfo_CustomColor_ViewMode() {
     LectureHeaderFields(
-        lecture = PreviewData.syllabusLecture,
+        lecture = LecturePreviewData.syllabusLecture,
         editMode = false,
         tableTheme = BuiltInTheme.SNUTT,
         onCourseTitleChange = {},
@@ -25,7 +25,7 @@ fun LectureHeaderFields_LectureUIInfo_CustomColor_ViewMode() {
 @Composable
 fun LectureHeaderFields_LectureUIInfo_CustomColor_EditMode() {
     LectureHeaderFields(
-        lecture = PreviewData.syllabusLecture,
+        lecture = LecturePreviewData.syllabusLecture,
         editMode = true,
         tableTheme = BuiltInTheme.SNUTT,
         onCourseTitleChange = {},
@@ -39,7 +39,7 @@ fun LectureHeaderFields_LectureUIInfo_CustomColor_EditMode() {
 @Composable
 fun LectureHeaderFields_LectureUIInfo_BuiltInColor_ViewMode() {
     LectureHeaderFields(
-        lecture = PreviewData.builtInColorLecture,
+        lecture = LecturePreviewData.builtInColorLecture,
         editMode = false,
         tableTheme = BuiltInTheme.SNUTT,
         onCourseTitleChange = {},
@@ -53,7 +53,7 @@ fun LectureHeaderFields_LectureUIInfo_BuiltInColor_ViewMode() {
 @Composable
 fun LectureHeaderFields_LectureUIInfo_BuiltInColor_EditMode() {
     LectureHeaderFields(
-        lecture = PreviewData.builtInColorLecture,
+        lecture = LecturePreviewData.builtInColorLecture,
         editMode = true,
         tableTheme = BuiltInTheme.SNUTT,
         onCourseTitleChange = {},
@@ -67,7 +67,7 @@ fun LectureHeaderFields_LectureUIInfo_BuiltInColor_EditMode() {
 @Composable
 fun LectureHeaderFields_NonUIInfo_ViewMode() {
     LectureHeaderFields(
-        lecture = PreviewData.searchedLecture,
+        lecture = LecturePreviewData.searchedLecture,
         editMode = false,
         tableTheme = BuiltInTheme.SNUTT,
         onCourseTitleChange = {},
