+++ app/src/screenshotTest/java/com/wafflestudio/snutt2/feature/lecturedetail/LectureSessionSectionScreenshotTest.kt
@@ -3,14 +3,14 @@ package com.wafflestudio.snutt2.feature.lecturedetail
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
 import com.android.tools.screenshot.PreviewTest
-import com.wafflestudio.snutt2.ui.preview.PreviewData
+import com.wafflestudio.snutt2.ui.preview.LecturePreviewData
 
 @PreviewTest
 @Preview(showBackground = true, widthDp = 360, locale = "ko")
 @Composable
 fun LectureSessionListSection_NonEmpty_ViewMode() {
     LectureSessionListSection(
-        sessions = PreviewData.syllabusLecture.lectureSessions,
+        sessions = LecturePreviewData.syllabusLecture.lectureSessions,
         editMode = false,
         onEditTime = { _, _ -> },
         onLocationChange = { _, _ -> },
@@ -24,7 +24,7 @@ fun LectureSessionListSection_NonEmpty_ViewMode() {
 @Composable
 fun LectureSessionListSection_NonEmpty_EditMode() {
     LectureSessionListSection(
-        sessions = PreviewData.syllabusLecture.lectureSessions,
+        sessions = LecturePreviewData.syllabusLecture.lectureSessions,
         editMode = true,
         onEditTime = { _, _ -> },
         onLocationChange = { _, _ -> },
